+++ compiler-plugin/src/main/kotlin/tech/mappie/MappieCompilerPluginRegistrar.kt
@@ -18,7 +18,6 @@ import tech.mappie.MappieCommandLineProcessor.Companion.ARGUMENT_REPORT_DIR
 import tech.mappie.MappieCommandLineProcessor.Companion.ARGUMENT_REPORT_ENABLED
 import tech.mappie.MappieCommandLineProcessor.Companion.ARGUMENT_STRICTNESS_JAVA_NULLABILITY
 import tech.mappie.MappieCommandLineProcessor.Companion.ARGUMENT_USE_DEFAULT_ARGUMENTS
-import tech.mappie.MappieCommandLineProcessor.Companion.ARGUMENT_WARNINGS_AS_ERRORS
 import tech.mappie.compiler_plugin.BuildConfig
 import tech.mappie.config.MappieConfiguration
 import tech.mappie.config.MappieModule
@@ -47,7 +46,6 @@ class MappieCompilerPluginRegistrar : CompilerPluginRegistrar() {
                 }
             },
             isMappieDebugMode = configuration.isStartedWithDependency(TESTUTIL_REGEX),
-            warningsAsErrors = configuration[ARGUMENT_WARNINGS_AS_ERRORS, false],
             useDefaultArguments = configuration[ARGUMENT_USE_DEFAULT_ARGUMENTS, true],
             namingConvention = configuration[ARGUMENT_NAMING_CONVENTION]?.let { NamingConventionMode.valueOf(it) } ?: NamingConventionMode.STRICT,
             strictEnums = configuration[ARGUMENT_STRICTNESS_ENUMS, true],