+++ compiler-plugin/src/main/kotlin/tech/mappie/config/MappieConfiguration.kt
@@ -8,7 +8,6 @@ data class MappieConfiguration(
     val configuration: CompilerConfiguration,
     val modules: EnumSet<MappieModule>,
     val isMappieDebugMode: Boolean,
-    val warningsAsErrors: Boolean,
     val useDefaultArguments: Boolean,
     val namingConvention: NamingConventionMode,
     val strictEnums: Boolean,